[Patchset 5] trace_processor: refactor to an ever present id approach
Instead of conditionally having the ID of the table based on whether or
not an arg is present for it, always include it in every table (just the
counters table for now but in future all tables).

This is also useful for future work on the raw table.

+++ src/trace_processor/trace_storage.h
@@ -379,7 +379,7 @@ class TraceStorage {
   }
 
   static RowId CreateRowId(TableId table, uint32_t row) {
-    static constexpr uint64_t kRowIdTableShift = 32ul;
+    static constexpr uint8_t kRowIdTableShift = 32;
     return (static_cast<uint64_t>(table) << kRowIdTableShift) | row;
   }
